+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 8.0
 - CSS minification optimizations in stylesheet.css & stylesheet_wkhtmltopdf.css
 - JS uglify optimizations in plugins.min.js & plugins.min.js.map
 - Upgrade Chart.js from 2.9.3 to 3.4.1 & save 40KB in chart.min.js & Charts.fnc.php
+- Fix "The gradebook configuration has been modified." note appearing twice in Grades/Configuration.php
 
 Changes in 7.9.1
 ----------------