+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 8.0
 - Upgrade Chart.js from 2.9.3 to 3.4.1 & save 40KB in chart.min.js & Charts.fnc.php
 - Fix "The gradebook configuration has been modified." note appearing twice in Grades/Configuration.php
 - Add warning in case all Assignments in Type have 0 Points (Extra Credit) in Assignments.php
+- Update French & Spanish translations in rosariosis.po
 
 Changes in 7.9.1
 ----------------